Resque: Don't push to Redis if json_encode fails.
The `Resque::push` method uses `json_encode` and doesn't check its
return value. This allows `false` or an empty string to be `rpush`ed
into Redis and the developer is none the wiser.

One such instance where `json_encode` will fail is if the arguments for
a job contain a string which is not properly UTF-8 encoded.